Add simple service module; move before_action logic from controller to it
