+++ app/controllers/services_controller.rb
@@ -1,6 +1,5 @@
 class ServicesController < ApplicationController
-  before_action :set_service, only: [:show, :edit, :update, :destroy]
-  before_action :require_permission, only: [:edit, :update, :destroy]
+  include ServicesService
   layout 'user' # TODO change to services new own layout
 
   # GET /services
@@ -12,6 +11,7 @@ class ServicesController < ApplicationController
   # GET /services/1
   # GET /services/1.json
   def show
+    set_service
   end
 
   # GET /services/new
@@ -22,6 +22,8 @@ class ServicesController < ApplicationController
 
   # GET /services/1/edit
   def edit
+    set_service
+    require_permission(@provider)
   end
 
   # POST /services
@@ -45,6 +47,8 @@ class ServicesController < ApplicationController
   # PATCH/PUT /services/1
   # PATCH/PUT /services/1.json
   def update
+    set_service
+    require_permission(@provider)
     @service.servicedata = params[:service][:servicedata]
     respond_to do |format|
       if @service.update(service_params)
@@ -60,6 +64,8 @@ class ServicesController < ApplicationController
   # DELETE /services/1
   # DELETE /services/1.json
   def destroy
+    set_service
+    require_permission(@provider)
     @service.destroy
     respond_to do |format|
       format.html { redirect_to edit_services_path(@provider), notice: 'Service was successfully destroyed.' }
@@ -68,18 +74,6 @@ class ServicesController < ApplicationController
   end
 
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_service
-      @service = Service.find(params[:id])
-      @provider = User.preload(:profile).find(@service.user_id)
-    end
-
-    def require_permission
-      if @provider.id != current_user.id
-        redirect_to user_path(current_user),
-                    notice: 'Editing or deleting of other\'s service is forbidden.'
-      end
-    end
 
     # Never trust parameters from the scary internet, only allow the white list through.
     def service_params
